+++ src/main/java/me/srrapero720/watermedia/api/network/patchs/ImgurPatch.java
@@ -1,12 +1,10 @@
-package me.srrapero720.watermedia.api.network.patches;
+package me.srrapero720.watermedia.api.network.patchs;
 
+import com.google.gson.annotations.SerializedName;
 import me.srrapero720.watermedia.api.MediaContext;
 import me.srrapero720.watermedia.api.MediaType;
 import me.srrapero720.watermedia.api.network.MediaURI;
 import me.srrapero720.watermedia.api.network.URIPatchException;
-import me.srrapero720.watermedia.api.network.patches.models.imgur.Gallery;
-import me.srrapero720.watermedia.api.network.patches.models.imgur.Image;
-import me.srrapero720.watermedia.api.network.patches.models.imgur.Response;
 import me.srrapero720.watermedia.tools.DataTool;
 import me.srrapero720.watermedia.tools.NetTool;
 
@@ -15,6 +13,8 @@ import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 
 public class ImgurPatch extends AbstractPatch {
     private static final String API_URL = "https://api.imgur.com/3";
@@ -130,5 +130,179 @@ public class ImgurPatch extends AbstractPatch {
         } finally {
             conn.disconnect();
         }
+    }
+
+    public static class Gallery {
+
+        @SerializedName("id")
+        public String id;
+
+        @SerializedName("title")
+        public String title;
+
+        @SerializedName("description")
+        public String desc;
+
+        @SerializedName("datetime")
+        public long date;
+
+        @SerializedName("cover")
+        public String cover;
+
+        @SerializedName("account_url")
+        public String accountUrl;
+
+        @SerializedName("account_id")
+        public int accountId;
+
+        @SerializedName("privacy")
+        public String privacy;
+
+        @SerializedName("layout")
+        public String layout;
+
+        @SerializedName("views")
+        public int views;
+
+        @SerializedName("link")
+        public String link;
+
+        @SerializedName("ups")
+        public int ups;
+
+        @SerializedName("downs")
+        public int downs;
+
+        @SerializedName("points")
+        public int points;
+
+        @SerializedName("score")
+        public int score;
+
+        @SerializedName("is_album")
+        public boolean album;
+
+        @SerializedName("vote")
+        public String vote;
+
+        @SerializedName("comment_count")
+        public int commentCount;
+
+        @SerializedName("images_count")
+        public int imagesCount;
+
+        @SerializedName("images")
+        public Image[] images;
+
+        public String getFormattedDate() {
+            Date time = new Date(date * 1000);
+            return new java.text.SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(time);
+        }
+    }
+
+    public static class Image {
+
+        @SerializedName("id")
+        public String id;
+
+        @SerializedName("deletehash")
+        public String deleteHash;
+
+        @SerializedName("account_id")
+        public int accountId;
+
+        @SerializedName("account_url")
+        public String accountUrl;
+
+        @SerializedName("ad_type")
+        public int adType;
+
+        @SerializedName("ad_url")
+        public String adUrl;
+
+        @SerializedName("title")
+        public String title;
+
+        @SerializedName("description")
+        public String desc;
+
+        @SerializedName("name")
+        public String name;
+
+        @SerializedName("type")
+        public String type;
+
+        @SerializedName("width")
+        public int width;
+
+        @SerializedName("height")
+        public int height;
+
+        @SerializedName("size")
+        public int size;
+
+        @SerializedName("views")
+        public int views;
+
+        @SerializedName("section")
+        public String section;
+
+        @SerializedName("vote")
+        public String vote;
+
+        @SerializedName("bandwidth")
+        public long bandwidth;
+
+        @SerializedName("animated")
+        public boolean anim;
+
+        @SerializedName("favorite")
+        public boolean fav;
+
+        @SerializedName("in_gallery")
+        public boolean gallery;
+
+        @SerializedName("in_most_viral")
+        public boolean viral;
+
+        @SerializedName("has_sound")
+        public boolean sound;
+
+        @SerializedName("is_ad")
+        public boolean ad;
+
+        @SerializedName("nsfw")
+        public String nsfw;
+
+        @SerializedName("link")
+        public String link;
+
+    //    @SerializedName("tags")
+    //    public Object[] tags;
+
+        @SerializedName("datetime")
+        public long date;
+
+        @SerializedName("mp4")
+        public String mp4;
+
+        @SerializedName("hls")
+        public String hls;
+
+        public String getFormattedDate() {
+            Date time = new Date(date * 1000);
+            return new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(time);
+        }
+    }
+
+    public static class Response<T> {
+        @SerializedName("data")
+        public T data;
+
+        @SerializedName("success")
+        public boolean success;
+
+        @SerializedName("status")
+        public int status;
     }
 }
