--- src/test/java/me/srrapero720/watermedia/tests/network/NetworkFixersTest.java
@@ -1,6 +1,0 @@
-package me.srrapero720.watermedia.tests.network;
-
-public class NetworkFixersTest {
-
-
-}
