+++ src/test/java/org/watermedia/BootstrapTest.java
@@ -1,10 +1,13 @@
-package me.srrapero720.watermedia.tests.bootstrap;
+package org.watermedia;
 
 import me.srrapero720.watermedia.WaterMedia;
+import me.srrapero720.watermedia.api.MediaContext;
 import me.srrapero720.watermedia.loader.ILoader;
 import org.junit.jupiter.api.BeforeAll;
 
+
 public class BootstrapTest {
+    public static final MediaContext CONTEXT = new MediaContext.Simple("test", "Test");
 
     @BeforeAll
     public static void testBootstrap() throws Exception {
