+++ src/test/java/org/watermedia/Seven7ExtractorTest.java
@@ -1,4 +1,4 @@
-package me.srrapero720.watermedia.tests.compress;
+package org.watermedia;
 
 import me.srrapero720.watermedia.tools.IOTool;
 import org.apache.logging.log4j.Marker;
